% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_electoral.R
\name{circ_electoral}
\alias{circ_electoral}
\title{Función para obtener las circunscripciones electorales}
\usage{
circ_electoral(eleccion = "elecciones_constitucion")
}
\arguments{
\item{eleccion}{eleccion para la cual se desea obtener las circunscripciones
electorales}
}
\description{
Función para obtener las circunscripciones electorales
}
\examples{
circ_electoral('elecciones_constitucion')
circ_electoral('elecciones_convencion')
}
