% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesas_circ_electoral.R
\name{mesas_circ_electoral}
\alias{mesas_circ_electoral}
\title{Función para obtener las mesas de votación de una circunscripción electoral}
\usage{
mesas_circ_electoral(
  eleccion = "elecciones_constitucion",
  id_circ_elect = "7479"
)
}
\arguments{
\item{eleccion}{elección para la cual se desea obtener las mesas}

\item{id_circ_elect}{id de circunscripción electoral}
}
\description{
Función para obtener las mesas de votación de una circunscripción electoral
}
\examples{
mesas_circ_electoral('elecciones_constitucion', 7479)

}
