% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_json.R
\name{obtener_json}
\alias{obtener_json}
\title{Función para obtener JSON desde url}
\usage{
obtener_json(url)
}
\arguments{
\item{url}{url de http://servelelecciones.cl correspondiente a un archivo JSON}
}
\description{
Función para obtener JSON desde url
}
\examples{
obtener_json('http://servelelecciones.cl/data/elecciones_constitucion/computomesas/70071112.json')
}
