% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultado_mesa.R
\name{resultado_mesa}
\alias{resultado_mesa}
\title{Función para obtener los resultados de una mesa de votación}
\usage{
resultado_mesa(eleccion = "elecciones_constitucion", id_mesa = "70071112")
}
\arguments{
\item{eleccion}{elección para la cual se desea obtener las circunscripciones
electorales}

\item{id_circ_elect}{id de circunscripción electoral}
}
\description{
Función para obtener los resultados de una mesa de votación
}
\examples{
mesas_circ_electoral('elecciones_constitucion', 7479)

}
